Moved math to seperate module
