+++ src/Math/Number.ts
