+++ src/QuantumComputing/Gates/QuantumGate.ts
@@ -1,4 +1,4 @@
-import Matrix from '../Math/Matrix';
+import Matrix from '../../Math/Matrix';
 
 export default class QuantumGate {
   public matrix: Matrix;
