+++ src/QuantumComputing/Gates/XGate.ts
@@ -1,6 +1,6 @@
-import ComplexNumber from '../Math/ComplexNumber';
-import Matrix from '../Math/Matrix';
-import RealNumber from '../Math/RealNumber';
+import ComplexNumber from '../../Math/ComplexNumber';
+import Matrix from '../../Math/Matrix';
+import RealNumber from '../../Math/RealNumber';
 import QuantumGate from './QuantumGate';
 
 export default class XGate extends QuantumGate {
