+++ src/QuantumComputing/Gates/YGate.ts
@@ -1,7 +1,7 @@
-import ComplexNumber from '../Math/ComplexNumber';
-import Matrix from '../Math/Matrix';
-import RealNumber from '../Math/RealNumber';
-import Vector from '../Math/Vector';
+import ComplexNumber from '../../Math/ComplexNumber';
+import Matrix from '../../Math/Matrix';
+import RealNumber from '../../Math/RealNumber';
+import Vector from '../../Math/Vector';
 import QuantumGate from './QuantumGate';
 
 export default class YGate extends QuantumGate {
